945: Rename GDrive setup to migration
