+++ app/views/gdrive/folders/show.html.erb
@@ -1,8 +1,8 @@
 <% content_for(:title) { "Google Drive" } %>
 <% content_for(:action_links) do %>
   <% if @auth_policy.index? %>
-    <%= link_to(icon_tag("gear") << " Setup",
-                gdrive_auth_url(host: Settings.url.host, community_id: current_community.id),
+    <%= link_to(icon_tag("file-import") << " Migration",
+                gdrive_migration_auth_url(host: Settings.url.host, community_id: current_community.id),
                 class: "btn btn-primary") %>
   <% end %>
 <% end %>
